+++ src/helper.h
@@ -5,7 +5,7 @@
 #include "raylib.h"
 #include "object.h"
 
-#define NO_SELECTION (ssize_t)-1
+#define NO_SELECTION (int16_t)-1
 #define NO_OBJECT NULL
 #define SELECTED_ID(id) (id != NO_SELECTION)
 #define SELECTED_OBJ(obj) (obj != NO_OBJECT)
@@ -27,13 +27,13 @@ Vector3 NearestPointOnLine(
 // sets the dynamic selection uniform to selected objects properties
 void object_dynamic_assignment(DynShader *shader, Object *obj);
 // append static (unselected, constant properties like pos, rot, etc.) object to shader ray march map
-const char *object_static_map_entry(Object *obj, size_t index);
+const char *object_static_map_entry(Object *obj, int16_t index);
 // append dynamnic (selected, uniform properties) object to shader ray march map
 const char *object_dynamic_map_entry(Object *obj);
 // create the entire (with all objects) object map
-DynShader object_map(DA *da, size_t selection, bool colour_index);
+DynShader object_map(DA *da, int16_t selection, bool colour_index);
 
 // index (colour) of pixel at `pos`
-size_t object_at_pos(Vector2 pos, Camera *camera, DA *objects);
+int16_t object_at_pos(Vector2 pos, Camera *camera, DA *objects);
 
 #endif // __HELPER_H__
